drm/msm/dsi-staging: fix lane idle check logic for dsi phy
Currently in DSI PHY v3 and v4, the check for lane idle state
expects the state of unused lanes bit to be zero, which is
not the case because if any particular lane is not used then
that lane state will be set to one which means stop state.
This causes ulps enable to fail as the lane idle check fails.
This change fixes the logic.

Change-Id: I9f2dcc4619ac119cef4e353524aebbfea19b2aba
Signed-off-by: Sandeep Panda <spanda@codeaurora.org>
Signed-off-by: Abhijith Desai <desaia@codeaurora.org>