arm64: Add padding to thread_info structure
Both flags and preempt_count variables of thread_info are falling
into the same cache line. When one of the cores is incrementing/
decrementing the preempt_count continuously in a tight loop with
spin_lock_irqsave() and spin_unlock_irqrestore() APIs, this causes the
flags update to be delayed. When the thread_info flags updates are
delayed, it leads to increased tasks scheduling latencies. So move the
thread_info's preempt_count into different cache line by adding padding.

Change-Id: I950d15a02b4df0c31862c3a11eb1a3d8fc17dff8
Signed-off-by: Prasad Sodagudi <psodagud@codeaurora.org>