Dynamic Stune Boost: add tunable to control resetting stune boost
This was added to circumvent the issue of conflicting unboosts between Pixel's powerhal and Dynamic
Stune Boost. If DSB decides to unboost while powerhal is boosted, it will lead to the device being
unboosted. This is bad, hence add a sysfs tunable to allow disabling unboosting on DSB's end and let
powerhal reset the value on its own.

Add it and enable it by default as it is the original intended behaviour, those affected can turn
it off if needed.

I personally don't really like this solution as it seems a little hacky, but this
is only temporary and a more robust solution will be implemented in the near
future.